+++ CHANGELOG.md
@@ -7,9 +7,16 @@
 * clj-reload
 * attribute :chatgpt is unnecessary. if found chatgpt's answer, show it.
 * admin/toggle-status
-* digest namespace
+* digest namespace. used only from answers.
 * time restriction
 * sometime, author is empty when sending comments.
+
+
+# 0.2.11-SNAPSHOT
+
+- added user/problem! and user/problems!
+- refactord
+- app.melt
 
 
 # 0.2.10 (2025-09-16)